About James Smithers
For most of my life I have worked in the service industry, I've always had a easy time speaking and relating to costumers and giving them the best experience possible. I have worked for a startup company for almost a year developing them into one of the most profitable firms in the retail sector. Also learning the ins and outs of how to run a business properly and efficiently. In the last few months, I have helped start an eclectic lawn care company, and get logistically prepared for our first working season. As a young kid I had the privilege to travel all over the world and I learned the value of hard work and dedication and how it relates to success. I would describe myself lately as an entrepreneur with outstanding knowledge of what it takes to truly see success. I’m a fantastic negotiator and have success at working efficiently under pressure, and adept at decision making and accurate forecasts.